[WIP] [Sprint/Moose] (goal): Fully implement recurring subscriptions (#470)
* (feat): Show recurring subscriptions on billing settings

* (chore): Move payments/subscriptions from v2 to v1

* (fix): USD not showing decimals on subscriptions

* (feat): Hide Cancel button selectively

* (feat): add checkbox for recurring wires when points is selected

* (feat): remove condition to subscription's cancel button

* (fix): use correct variables

+++ app/src/modules/settings/billing/subscriptions/subscriptions.component.ts
@@ -43,13 +43,17 @@ export class SettingsBillingSubscriptionsComponent {
   }
 
   cancel(i: number) {
+    if (!confirm('Are you sure you want to cancel this subscription?')) {
+      return;
+    }
+
     this.inProgress = true;
     this.cd.detectChanges();
 
     let subscription = this.subscriptions[i];
     this.client.delete(`api/v1/payments/subscriptions/${subscription.id}`)
       .then(() => {
-        this.subscriptions.splice(i, 0);
+        this.subscriptions.splice(i, 1);
         this.inProgress = false;
         this.cd.detectChanges();
       })
